split up composer into separate files
