+++ apps/expo/src/app/(tabs)/(feeds,search,notifications,self)/profile/[handle]/post/[id].tsx
@@ -25,8 +25,8 @@ import { PrimaryPost } from "~/components/primary-post";
 import { QueryWithoutData } from "~/components/query-without-data";
 import { Text as ThemedText } from "~/components/text";
 import { useAgent } from "~/lib/agent";
+import { useComposer } from "~/lib/composer/utils";
 import { useTabPressScroll } from "~/lib/hooks";
-import { useComposer } from "~/lib/hooks/composer";
 import { useContentFilter, type FilterResult } from "~/lib/hooks/preferences";
 import { assert } from "~/lib/utils/assert";
 import { useUserRefresh } from "~/lib/utils/query";
